requirements.txt
dash~=2.8.1
dash_bootstrap_components~=1.4.0
gunicorn~=20.1.0
pandas~=1.5.3
altair_data_server~=0.4.1
numpy~=1.24.1
Pillow~=9.4.0